Change MasteryLevel fields to int, participantId type to long

+++ src/main/kotlin/org/ratifire/matcherservice/entity/ParticipantEntity.kt
@@ -9,7 +9,7 @@ import java.util.*
 @Document("Participant")
 data class ParticipantEntity(
     @Id val id: ObjectId? = null,
-    val participantId: Int,
+    val participantId: Long,
     val specialization: String,
     val type: ParticipantType,
     val masteryLevel: Int,
